fix(functional): satisfy teardown review gates
