+++ tests/functional/sessions/chat_sessions/root_composition/acp_server_composition_test.go
@@ -1,3 +1,4 @@
+// Functional owner: sessions/chat_sessions/root_composition.
 package root_composition_test
 
 import (
