+++ tests/functional/sessions/chat_sessions/root_composition/resource_census_test.go
@@ -1,3 +1,4 @@
+// Functional owner: sessions/chat_sessions/root_composition.
 package root_composition_test
 
 import (
